requirements.txt
asgiref==3.8.1
Django==5.0.4
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-js-asset==2.2.0
django-rest-framework==0.1.0
djangorestframework==3.15.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
